avi now can save audio (mme_saveWav 2), fixed demo anim feature a bit with Scooper's hand (setting correct children angles is left2do), cleaned code, fixed scoreboard elements, fixed 2d under water, added force lightning looping sound support from SP, fixed colours of version in the console, forcing core affinity feature now supports flags, fixed lacking memory in sound code

+++ codemp/client/FxPrimitives.h
@@ -590,7 +590,7 @@ public:
 	vec2_t	mST[MAX_CPOLY_VERTS];
 
 	float	mRot[3][3];
-	int		mLastFrameTime;
+	float	mLastFrameTime;
 
 
 	CPoly() {}
